Move message processing related classes into a new package
