+++ src/main/java/io/github/ust/mico/kafkafaasconnector/MessageListener.java
@@ -20,6 +20,8 @@
 package io.github.ust.mico.kafkafaasconnector;
 
 import com.fasterxml.jackson.databind.JsonNode;
+import io.github.ust.mico.kafkafaasconnector.MessageProcessing.FaasController;
+import io.github.ust.mico.kafkafaasconnector.MessageProcessing.KafkaMessageSender;
 import io.github.ust.mico.kafkafaasconnector.configuration.KafkaConfig;
 import io.github.ust.mico.kafkafaasconnector.exception.MicoCloudEventException;
 import io.github.ust.mico.kafkafaasconnector.kafka.MicoCloudEventImpl;
