+++ src/test/java/io/github/ust/mico/kafkafaasconnector/MessageListenerTests.java
@@ -21,6 +21,9 @@ package io.github.ust.mico.kafkafaasconnector;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import io.cloudevents.json.Json;
+import io.github.ust.mico.kafkafaasconnector.MessageProcessing.CloudEventManipulator;
+import io.github.ust.mico.kafkafaasconnector.MessageProcessing.FaasController;
+import io.github.ust.mico.kafkafaasconnector.MessageProcessing.KafkaMessageSender;
 import io.github.ust.mico.kafkafaasconnector.configuration.KafkaConfig;
 import io.github.ust.mico.kafkafaasconnector.exception.MicoCloudEventException;
 import io.github.ust.mico.kafkafaasconnector.kafka.MicoCloudEventImpl;
